+++ src/commands/noreact.ts
@@ -7,7 +7,7 @@ import {
 } from 'discord.js'
 import { getNoReactChannels, setNoReactChannels } from '../data'
 import { CommandMap } from '../types'
-import { defineCommand } from '../util'
+import { defineCommand } from '../util/guild'
 
 export default defineCommand({
     data: new SlashCommandBuilder()
